--- a/docs/src/api.md
+++ b/docs/src/api.md
@@ -40,7 +40,7 @@ even though [`Prior()`](@ref) is actually defined in the `Turing.Inference` modu
 | `@model`             | [`DynamicPPL.@model`](@extref)             | Define a probabilistic model                                                                 |
 | `@varname`           | [`AbstractPPL.@varname`](@extref)          | Generate a `VarName` from a Julia expression                                                 |
 | `to_submodel`        | [`DynamicPPL.to_submodel`](@extref)        | Define a submodel                                                                            |
-| `prefix`             | [`DynamicPPL.prefix`](@extref)             | Prefix all variable names in a model with a given symbol                                     |
+| `prefix`             | [`DynamicPPL.prefix`](@extref)             | Prefix all variable names in a model with a given VarName                                    |
 | `LogDensityFunction` | [`DynamicPPL.LogDensityFunction`](@extref) | A struct containing all information about how to evaluate a model. Mostly for advanced users |
 
 ### Inference