This PR adds compatibility with DynamicPPL 0.36.

Given that there are some reasonably breaking changes in the way that submodels behave, I've decided to increment the minor version number for Turing.

Because submodels now also prefix VarNames in a 'proper' way, submodel variables now have non-identity lenses. In order to allow Gibbs to sample models with submodels, this PR also fixes #2403 by letting GibbsContext store a tuple of VarNames rather than just Symbols. Performance is compared below.

There's some nasty lack of reproducibility in HMCDA sampling, which is causing the failures in `test/mcmc/Inference`. I haven't yet been able to repro this locally. However, I would like to deal with this in a separate PR, as the failure isn't introduced by this PR (the last CI run on main was already failing).

This PR supersedes the CompatHelper ones at #2501 #2532 #2533 

## Benchmarks

I'm not 100% convinced of these, because my benchmarks have returned wildly variable results upon being run multiple times / when I restart my Julia session. Sometimes when I run `sample(rng, model, spl, 10000)` it runs in < 1 second and sometimes it takes 4 seconds. (This is after running it multiple times in the same session, so it's not an issue of JIT compilation.) I think I would probably feel a little bit more reassured if somebody else could test them out.

However, broadly speaking, I think it's safe to say that I don't have any evidence of performance getting _worse_. Another noteworthy point is that CI for `test/mcmc/gibbs` (on ubuntu-latest, 1 thread) for this PR ran in 1 hr 25 min, which is the same time it took for the same job on `main`. So it seems unlikely that there are any serious performance implications (good or bad).

```julia
using AbstractMCMC: step
using DynamicPPL, Turing
using StableRNGs: StableRNG
using Chairmarks

@model function f()
    x ~ Normal()
    y ~ Normal()
end
rng = StableRNG(468)
model = f()
spl = Gibbs(@varname(x) => MH(), @varname(y) => MH())
dspl = DynamicPPL.Sampler(spl)

# This benchmark compares the initial step of MCMC
@be step($rng, $model, $dspl)

# Before: 91.250 μs
# After: 91.208 µs

# This compares the 2nd step onwards
_, st = step(rng, model, dspl)
@be step($rng, $model, $dspl, $st)

# Before: 66.500 μs
# After: 66.417 μs
```